containers/El-Campello/telebot/requirements.txt
python-telegram-bot
requests
numpy
telebot
